Refactor tool execution flow
GitlabClient and GitlabService now support pagination for all APIs.
GitService not offers two main operations, (1) clone or init submodules if clone fails, (2) only clone repository.
The command implementation searches for the group, retrieves it (by blocking) and reports if the group is not found. If the group is found then the flow continues to fetching the group projects.